refactor: music system
