+++ commands/playlist.js
@@ -1,20 +1,8 @@
-import {
-  createAudioPlayer,
-  getVoiceConnection,
-  joinVoiceChannel,
-  NoSubscriberBehavior
-} from "@discordjs/voice";
 import { MessageEmbed } from "discord.js";
-import YouTubeAPI from "simple-youtube-api";
-import Scdl from "soundcloud-downloader";
-import { play } from "../include/play.js";
-import { generateQueue } from "../utils/queue.js";
-import { config } from "../utils/config.js";
+import { getPlaylist } from "../music/getPlaylist.js";
+import { startQueue } from "../music/startQueue.js";
 import { i18n } from "../utils/i18n.js";
-
-const { MAX_PLAYLIST_SIZE, SOUNDCLOUD_CLIENT_ID, YOUTUBE_API_KEY } = config;
-const youtube = new YouTubeAPI(YOUTUBE_API_KEY);
-const scdl = Scdl.create();
+import { generateQueue } from "../utils/queue.js";
 
 export default {
   name: "playlist",
@@ -24,7 +12,7 @@ export default {
   permissions: ["CONNECT", "SPEAK", "ADD_REACTIONS", "MANAGE_MESSAGES"],
   async execute(message, args) {
     const { channel } = message.member.voice;
-    const serverQueue = message.client.queue.get(message.guild.id);
+    const queue = message.client.queue.get(message.guild.id);
 
     if (!args.length)
       return message
@@ -33,65 +21,27 @@ export default {
 
     if (!channel) return message.reply(i18n.__("playlist.errorNotChannel")).catch(console.error);
 
-    if (serverQueue && channel.id !== serverQueue.channel.id)
+    if (queue && channel.id !== queue.channel.id)
       return message
         .reply(i18n.__mf("play.errorNotInSameChannel", { user: message.client.user.username }))
         .catch(console.error);
 
-    const search = args.join(" ");
-    const pattern = /^.*(youtu.be\/|list=)([^#\&\?]*).*/gi;
-    const url = args[0];
-    const urlValid = pattern.test(args[0]);
+    const { playlist, videos } = await getPlaylist({ message, args });
 
-    const queueConstruct = generateQueue(message.channel, channel);
-
-    let playlist = null;
-    let videos = [];
-
-    if (urlValid) {
-      try {
-        playlist = await youtube.getPlaylist(url, { part: "snippet" });
-        videos = await playlist.getVideos(MAX_PLAYLIST_SIZE || 10, { part: "snippet" });
-      } catch (error) {
-        console.error(error);
-        return message.reply(i18n.__("playlist.errorNotFoundPlaylist")).catch(console.error);
-      }
-    } else if (scdl.isValidUrl(args[0])) {
-      if (args[0].includes("/sets/")) {
-        message.reply(i18n.__("playlist.fetchingPlaylist"));
-        playlist = await scdl.getSetInfo(args[0], SOUNDCLOUD_CLIENT_ID);
-        videos = playlist.tracks.map((track) => ({
-          title: track.title,
-          url: track.permalink_url,
-          duration: track.duration / 1000
-        }));
-      }
+    if (queue) {
+      queue.songs.push(...videos);
     } else {
-      try {
-        const results = await youtube.searchPlaylists(search, 1, { part: "id" });
-        playlist = results[0];
-        videos = await playlist.getVideos(MAX_PLAYLIST_SIZE, { part: "snippet" });
-      } catch (error) {
-        console.error(error);
-        return message.reply(error.message).catch(console.error);
-      }
+      const queueConstruct = generateQueue(message.channel, channel);
+      queueConstruct.songs.push(...videos);
+
+      message.client.queue.set(message.guild.id, queueConstruct);
+
+      startQueue({ message, channel });
     }
-
-    const newSongs = videos
-      .filter((video) => video.title != "Private video" && video.title != "Deleted video")
-      .map((video) => {
-        return {
-          title: video.title,
-          url: video.url,
-          duration: video.durationSeconds
-        };
-      });
-
-    serverQueue ? serverQueue.songs.push(...newSongs) : queueConstruct.songs.push(...newSongs);
 
     let playlistEmbed = new MessageEmbed()
       .setTitle(`${playlist.title}`)
-      .setDescription(newSongs.map((song, index) => `${index + 1}. ${song.title}`).join("\n"))
+      .setDescription(videos.map((song, index) => `${index + 1}. ${song.title}`).join("\n"))
       .setURL(playlist.url)
       .setColor("#F8AA2A")
       .setTimestamp();
@@ -106,32 +56,5 @@ export default {
         embeds: [playlistEmbed]
       })
       .catch(console.error);
-
-    if (!serverQueue) {
-      message.client.queue.set(message.guild.id, queueConstruct);
-
-      try {
-        queueConstruct.player = createAudioPlayer({
-          behaviors: {
-            noSubscriber: NoSubscriberBehavior.Pause
-          }
-        });
-
-        queueConstruct.connection = joinVoiceChannel({
-          channelId: channel.id,
-          guildId: channel.guild.id,
-          adapterCreator: channel.guild.voiceAdapterCreator
-        });
-
-        play(queueConstruct.songs[0], message);
-      } catch (error) {
-        console.error(error);
-        message.client.queue.delete(message.guild.id);
-
-        getVoiceConnection(channel.guild.id).destroy();
-
-        return message.reply(i18n.__mf("play.cantJoinChannel", { error: error })).catch(console.error);
-      }
-    }
   }
 };
